Make action_mailer_tls compatible with 1.8.7.
In most cases you should use the built in support in Ruby 1.8.7 and
Rails >= 2.2.1. But if you're upgrading an older app or are deploying to
an environment that differs from your current one it would be nice if
ActionMailerTLS didn't blow up. This way you can remove ActionMailerTLS
at your convenience.